debug:section:.text.entry: rename the section name of _entry
It is found that the gdb cannot pick the entry.S if its section is name
as .entry.text, so we rename its section name to be .text.entry, which
allows the gdb to pick it and add its to the .debug* sections. You can
check by objdump:
$riscv64-linux-gnu-objdump -g safeos.elf | grep entry
The command will show that entry.S will be add to the File Name Talbe if
the name _entry is .text.entry.

Signed-off-by: shijun zhao <zqyzsj@gmail.com>