Kmem.rs: Add a Kmem module to allocate physical memory
Add a kmem.rs file containing a Kmem struct, which can be created using the function new().
Provides two memory management interfaces palloc() and pfree().
The input of palloc() is usize and the output is Option<usize>.
The input to pfree() is the starting address and size of the free area.

Signed-off-by: haodi zhang <zhanghaodi@iie.ac.cn>.