compile-script:change the compile-script with multi-level Makefile
we changed the hierachy of Makefile into multi-level Makefile so t
hat we can control the project more carefully. we turn the compila
tion of c files int an archive file so that other rust crates can
link to it with just cargo in cast.

Signed-off:	daibo <d1716911340@gmail.com>