cross-crate-static: add 'cross-crate' for cross-crate-static
we add the cross-crate-static crate for those who want to expose t
heir functions to others. If they want to do like that they need t
o use it as  their dependency, and rust_main should use them as it
s dependencies. However, if we want to use the linkme, we mustn't
use the native c linker but the rust doing  linking itself.

Signed-off:     daibo <d1716911340@gmail.com>