cross-crate-test:add test_server and test_client crate for test
we add test_server crate and test_client crate for the linkme test.
and add this test to the qemu-test compilation target. however, to
use linkme, we mustn't use the native c 's linker, but rust linking
itself. If not, the linkme's section will not be loaded into the elf
file.

Signed-off: 	daibo <d1716911340@gmail.com>